+++ src/main/java/com/harusari/chainware/purchase/query/dto/PurchaseOrderSearchCondition.java
@@ -1,4 +1,4 @@
-package com.harusari.chainware.purchase.query.dto.request;
+package com.harusari.chainware.purchase.query.dto;
 
 import lombok.Getter;
 import lombok.Setter;
