+++ src/main/java/com/harusari/chainware/purchase/query/mapper/PurchaseOrderQueryMapper.java
@@ -1,9 +1,9 @@
 package com.harusari.chainware.purchase.query.mapper;
 
-import com.harusari.chainware.purchase.query.dto.request.PurchaseOrderSearchCondition;
-import com.harusari.chainware.purchase.query.dto.response.PurchaseOrderDetailResponse;
-import com.harusari.chainware.purchase.query.dto.response.PurchaseOrderProductResponse;
-import com.harusari.chainware.purchase.query.dto.response.PurchaseOrderSummaryResponse;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderSearchCondition;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderDetailResponse;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderProductResponse;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderSummaryResponse;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
