+++ src/main/java/com/harusari/chainware/purchase/query/service/PurchaseOrderQueryServiceImpl.java
@@ -1,11 +1,12 @@
 package com.harusari.chainware.purchase.query.service;
 
-import com.harusari.chainware.purchase.query.dto.request.PurchaseOrderSearchCondition;
-import com.harusari.chainware.purchase.query.dto.response.PurchaseOrderDetailResponse;
-import com.harusari.chainware.purchase.query.dto.response.PurchaseOrderSummaryResponse;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderSearchCondition;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderDetailResponse;
+import com.harusari.chainware.purchase.query.dto.PurchaseOrderSummaryResponse;
 import com.harusari.chainware.purchase.query.mapper.PurchaseOrderQueryMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -16,11 +17,13 @@ public class PurchaseOrderQueryServiceImpl implements PurchaseOrderQueryService 
     private final PurchaseOrderQueryMapper mapper;
 
     @Override
+    @Transactional
     public List<PurchaseOrderSummaryResponse> getPurchaseOrders(Long memberId, PurchaseOrderSearchCondition condition) {
         return mapper.findPurchaseOrders(memberId, condition);
     }
 
     @Override
+    @Transactional
     public PurchaseOrderDetailResponse getPurchaseOrderDetail(Long memberId, Long purchaseOrderId) {
         var order = mapper.findPurchaseOrderById(memberId, purchaseOrderId);
         var products = mapper.findProductsByPurchaseOrderId(memberId, purchaseOrderId);
@@ -36,4 +39,5 @@ public class PurchaseOrderQueryServiceImpl implements PurchaseOrderQueryService 
 
         return response;
     }
+
 }
